+++ test/life-cycle/save-mock-fs.ts
@@ -112,6 +112,7 @@ export function mockModules(): void {
       // make log4js and fs-extra happy in mocked env
       realpath: jest.fn(),
       promises: {
+        mkdir: jest.fn(),
         writeFile: jest.fn(),
         utimes: jest.fn(),
       },
@@ -120,14 +121,6 @@ export function mockModules(): void {
     fs.default = fs;
     return fs;
   });
-  jest.mock('mkdirp', () => ({
-    default: jest.fn(),
-    mkdirp: jest.fn(),
-  }));
-  jest.unstable_mockModule('mkdirp', () => ({
-    default: jest.fn(),
-    mkdirp: jest.fn(),
-  }));
   jest.mock('log4js', () => ({
     configure: jest.fn(),
     getLogger: jest.fn(),